refactor: address code review findings on transformation pipelines
- extract project_service.log_transformations_or_restore, the compensating
  write both the transform endpoint and the pipeline apply endpoint used as
  a verbatim copy; a partial log now restores the file from one place
- move _get_owned_pipeline into dependencies.fetch_owned_pipeline, next to
  fetch_owned_project it mirrors
- pipeline_service returns PipelineCompatibilityResponse instead of
  dicts shaped like it, so the type checks the shape
- captureStep() in transformFormProps replaces the pasted capture block in
  all 13 transform forms
- deletePipeline returns response.data like every other API function, and
  the tab uses getErrorMessage instead of a local reader that rendered a
  422 detail array as a React child
- add toStepInputs and stepFailureMessage for the draft-to-API mapping and
  the failure sentence, each of which was written twice
- split PipelinesTab (581 lines) into Section, StepText, HistoryList,
  DraftSequence, EligibilityCheck, SaveBar and PipelineCard; the tab is now
  composition, and panelNames holds the shared panel constant
- rename checkSteps to checkDraftPipelineSteps
- cover the extracted seams: log_transformations_or_restore, captureStep,
  toStepInputs and stepFailureMessage
