+++ dataloom-frontend/src/Components/forms/FillEmptyForm.tsx
@@ -1,4 +1,4 @@
-import type { TransformFormProps } from "./transformFormProps";
+import { captureStep, type TransformFormProps } from "./transformFormProps";
 import { useState, FormEvent } from "react";
 import { transformProject } from "../../api";
 import { useProjectContext } from "../../context/ProjectContext";
@@ -54,10 +54,7 @@ const FillEmptyForm = ({ projectId, onClose, onCapture }: TransformFormProps) =>
           fill_value: strategy === "custom" ? fillValue : null,
         },
       };
-      if (onCapture) {
-        onCapture({ action_type: payload.operation_type, action_details: payload });
-        return;
-      }
+      if (captureStep(onCapture, payload)) return;
 
       const response = await transformProject(projectId, payload, {
         preview: true,
