+++ dataloom-frontend/src/Components/forms/FilterForm.tsx
@@ -1,4 +1,4 @@
-import type { TransformFormProps } from "./transformFormProps";
+import { captureStep, type TransformFormProps } from "./transformFormProps";
 import { useState, ChangeEvent, FormEvent } from "react";
 import { transformProject } from "../../api";
 import { FILTER } from "../../constants/operationTypes";
@@ -57,10 +57,7 @@ const FilterForm = ({ projectId, onClose, onCapture }: TransformFormProps) => {
         operation_type: FILTER,
         parameters: filterParams,
       };
-      if (onCapture) {
-        onCapture({ action_type: payload.operation_type, action_details: payload });
-        return;
-      }
+      if (captureStep(onCapture, payload)) return;
 
       const response = await transformProject(projectId, payload, {
         preview: true,
