+++ dataloom-frontend/src/Components/forms/FormulaColumnForm.tsx
@@ -1,4 +1,4 @@
-import type { TransformFormProps } from "./transformFormProps";
+import { captureStep, type TransformFormProps } from "./transformFormProps";
 import { useState, FormEvent } from "react";
 import { transformProject } from "../../api";
 import { useProjectContext } from "../../context/ProjectContext";
@@ -40,10 +40,7 @@ const FormulaColumnForm = ({ projectId, onClose, onCapture }: TransformFormProps
           expression: expression.trim(),
         },
       };
-      if (onCapture) {
-        onCapture({ action_type: payload.operation_type, action_details: payload });
-        return;
-      }
+      if (captureStep(onCapture, payload)) return;
 
       const response = await transformProject(projectId, payload, {
         preview: true,
