+++ dataloom-frontend/src/Components/forms/GroupByForm.tsx
@@ -1,4 +1,4 @@
-import type { TransformFormProps } from "./transformFormProps";
+import { captureStep, type TransformFormProps } from "./transformFormProps";
 import { useState, useRef, useEffect, FormEvent } from "react";
 import { transformProject } from "../../api";
 import { GROUPBY } from "../../constants/operationTypes";
@@ -57,10 +57,7 @@ const GroupByForm = ({ projectId, onClose, onCapture }: TransformFormProps) => {
           agg_function: aggFunction,
         },
       };
-      if (onCapture) {
-        onCapture({ action_type: payload.operation_type, action_details: payload });
-        return;
-      }
+      if (captureStep(onCapture, payload)) return;
 
       const response = await transformProject(projectId, payload, {
         preview: true,
