+++ dataloom-frontend/src/Components/forms/PivotTableForm.tsx
@@ -1,4 +1,4 @@
-import type { TransformFormProps } from "./transformFormProps";
+import { captureStep, type TransformFormProps } from "./transformFormProps";
 import { useState, FormEvent } from "react";
 import useError from "../../hooks/useError";
 import { transformProject } from "../../api";
@@ -42,10 +42,7 @@ const PivotTableForm = ({ projectId, onClose, onCapture }: TransformFormProps) =
         operation_type: PIVOT_TABLES,
         pivot_query: { index: index.join(","), column, value, aggfun },
       };
-      if (onCapture) {
-        onCapture({ action_type: payload.operation_type, action_details: payload });
-        return;
-      }
+      if (captureStep(onCapture, payload)) return;
 
       const response = await transformProject(projectId, payload, {
         preview: true,
