+++ dataloom-frontend/src/Components/pipelines/panelNames.ts
@@ -1,0 +1,9 @@
+/**
+ * Docked panel names for the pipelines feature.
+ *
+ * Kept apart from the components so the tab and the feature registration can
+ * both name the same panel without importing each other.
+ */
+
+/** The step-builder panel the Pipelines tab opens for "add a step from scratch". */
+export const STEP_BUILDER_PANEL = "PipelineStepBuilder";
